Delete manifests/m/MicrosoftCorporation directory
I mistakenly added this package in MicrosoftCorporation directory. This directory was not present earlier. It was created when we added our package. Ideally, it's Microsoft package so it should be in Microsoft directory to maintain uniformity. I want to delete this directory in this PR and will raise another PR for adding package in Microsoft directory.